backend/requirements.txt
Authlib==0.14.3
fastapi==0.65.2
celery==5.0.5
redis==3.5.3
httpx==0.15.5
itsdangerous==1.1.0
Jinja2
psycopg2-binary==2.8.6
requests
SQLAlchemy==1.3.19
uvicorn==0.12.1
passlib==1.7.2
bcrypt==3.2.0
sqlalchemy-utils==0.36.8
python-multipart==0.0.5
pyjwt==1.7.1
alembic
python-dateutil
flower==1.1.0
statsmodels
influxdb-client==1.36.1

flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0; python_version < '3.8'
ipython~=8.10; python_version >= '3.8'
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0, <3.6.0
kedro~=0.18.8
kedro-telemetry~=0.2.0
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=7.2

python-dotenv
python-decouple